feat: Implement full technical audit recommendations
This commit addresses a comprehensive technical audit of the project, implementing a wide range of fixes and improvements to enhance security, performance, maintainability, and user experience.

Key changes include:

1.  **Security:**
    -   Patched a critical vulnerability in the admin rights check. The check is now performed server-side using a new `is_claims_admin()` Supabase RPC function that validates a JWT claim, replacing the insecure client-side check.
    -   All admin routes are now protected by a new `adminAuthMiddleware`.

2.  **Architecture & Refactoring:**
    -   Refactored the monolithic `server/index.js` into a clean, modular structure with controllers, services, routes, and middleware.
    -   Centralized Supabase client creation into a single `server/lib/supabaseClient.js` module.
    -   Centralized all API action strings into a single source of truth at `shared/constants.js`.

3.  **Performance & Optimization:**
    -   Optimized the Docker build process by implementing a multi-stage `Dockerfile` and a comprehensive `.dockerignore` file, resulting in a smaller and more secure production image.
    -   Replaced an inefficient user filtering method with a new `get_user_by_email` RPC function to fetch users directly from the database.

4.  **Bug Fixes & UX:**
    -   Fixed a bug in the client-side course search logic by using `flatMap` for robust searching in nested data structures.
    -   Improved the user notification message in the admin panel when a draft is created from a published course.

5.  **Testing:**
    -   Updated and fixed the entire test suite (`mocha`, `jsdom`) to work with the new, refactored, and authenticated application structure. All tests are now passing.
